# tests/contract/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOrValidationResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ValidationErrorResponse"
          }
        ]
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Structured errors emitted by application and cross-cutting middleware.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          },
          "error": {
            "anyOf": [
              {
                "const": "rate_limited",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "retry_after": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {
        "description": "A recursively typed JSON value; never an unconstrained Any schema.",
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ],
        "title": "JsonValue"
      },
...
      "UploadAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "aviso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aviso"
          },
          "estimated_bets": {
            "minimum": 0.0,
            "title": "Estimated Bets",
            "type": "integer"
          },
          "estimated_cost_usd": {
            "minimum": 0.0,
            "title": "Estimated Cost Usd",
            "type": "number"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "estimated_bets",
          "estimated_cost_usd",
          "status_url"
        ],
        "title": "UploadAcceptedResponse",
        "type": "object"
      },
      "UploadProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "failed": {
            "minimum": 0.0,
            "title": "Failed",
            "type": "integer"
          },
          "ignored": {
            "minimum": 0.0,
            "title": "Ignored",
            "type": "integer"
          },
          "over_limit": {
            "minimum": 0.0,
            "title": "Over Limit",
            "type": "integer"
          },
          "pending": {
            "minimum": 0.0,
            "title": "Pending",
            "type": "integer"
          },
          "percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent",
            "type": "integer"
          },
          "read": {
            "minimum": 0.0,
            "title": "Read",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "pending",
          "read",
          "failed",
          "ignored",
          "over_limit",
          "percent"
        ],
        "title": "UploadProgressResponse",
        "type": "object"
      },
      "UploadStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "bets_failed": {
            "minimum": 0.0,
            "title": "Bets Failed",
            "type": "integer"
          },
          "bets_processed": {
            "minimum": 0.0,
            "title": "Bets Processed",
            "type": "integer"
          },
          "concluido_em": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concluido Em"
          },
          "cost_usd": {
            "minimum": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "criado_em": {
            "format": "date-time",
            "title": "Criado Em",
            "type": "string"
          },
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          },
          "estimated_bets": {
            "minimum": 0.0,
            "title": "Estimated Bets",
            "type": "integer"
          },
          "estimated_cost_usd": {
            "minimum": 0.0,
            "title": "Estimated Cost Usd",
            "type": "number"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/UploadProgressResponse"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_messages": {
            "minimum": 0.0,
            "title": "Total Messages",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "filename",
          "total_messages",
          "estimated_bets",
          "estimated_cost_usd",
          "progress",
          "bets_processed",
          "bets_failed",
          "cost_usd",
          "erro",
          "criado_em",
          "concluido_em"
        ],
        "title": "UploadStatusResponse",
        "type": "object"
      },
      "ValidationErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueResponse"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
      "ValidationIssueResponse": {
        "additionalProperties": false,
        "properties": {
          "ctx": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctx"
          },
          "input": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssueResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CollectionToken": {
        "description": "Token opaco da extensão; armazenado no servidor somente como HMAC.",
        "in": "header",
        "name": "X-Coleta-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for authenticated betting, cash-flow, review, upload, collection, and dashboard workflows. The checked-in OpenAPI document is the compatibility contract.",
    "title": "Bancaemdia API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/upload": {
      "post": {
        "description": "Valida e guarda uma exportação do Telegram antes de agendar seu processamento assíncrono.",
        "operationId": "receber_export_api_v1_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "encoding": {
                "chat_filter": {
                  "explode": true,
                  "style": "form"
                }
              },
              "examples": {
                "default": {
                  "summary": "Exportação ZIP com filtro opcional",
                  "value": {
                    "chat_filter": [
                      123456789
                    ],
                    "file": "telegram-export.zip"
                  }
                }
              },
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "properties": {
                  "chat_filter": {
                    "description": "IDs de chats a aceitar dentro da exportação.",
                    "items": {
                      "type": "integer"
                    },
                    "type": "array"
                  },
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed multipart upload."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "example": "Content Too Large",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The upload exceeds either the declared application limit (JSON) or the server-wide streaming limit (plain text)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enviar exportação do Telegram"
      }
    },
    "/api/v1/upload/{job_id}": {
      "get": {
        "description": "Retorna estado, progresso, contagens e custo observado de um upload do usuário.",
        "operationId": "consultar_upload_api_v1_upload__job_id__get",
        "parameters": [
          {
            "description": "UUID público retornado quando o upload foi aceito.",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, expired, or invalid bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upload job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOrValidationResponse"
                }
              }
            },
            "description": "The request does not satisfy the published contract or a domain rule."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller exceeded an applicable request limit."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure with no sensitive exception details."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A required dependency is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consultar processamento do upload"
      }
    },
...
  }
}